requirements.txt
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
django-cors-headers==4.1.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.14
gunicorn==20.1.0